requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.1
future==1.0.0
iso8601==2.1.0
kiwisolver==1.4.9
matplotlib==3.10.6
numpy==2.3.3
packaging==25.0
pillow==11.3.0
pyparsing==3.2.5
pyserial==3.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
six==1.17.0
